refactor elemwise ops
